Drop PyPI configuration from repository
This should be configured on a user level and not commited to the
repository. It contains credentials for the PyPI account and thus
everyone could simply take over the account.

The proper place to store this configuration is:
* On Linux/OSX: `~/.pypirc`,
* On Windows: `%USERPROFILE%/.pypirc`.

In general, it is also recommended to even in that case not store the
password in the configuration file but pass it on upload or utilise
Twines keyring functionality.